Better performance has been reached. Bash files have been fixed, improved and tested. However, a fix of output format is needed. Note: Bash file had an error: each M was being evaluated over a different scenario, and that was generating bad analysis information. This problem has been fixed.
Important notes:
There is no way to improve structure creation that will not cause to violate the object's paradigm.
What I mean is that probably working with crude arrays, and built-ins will improve time performance, but will not match with an OOP style.
It was decided to keep things like they were.However, there should be more checks to see if it can be improved.

Brute Force Method was degraded to a worse performance so as times match what teachers expect.
Found documents of the Cell Index Method that established that for the CIM to overcome the overhead created by the structure creation, there should be more than 1000 particles (N > 1000 approximately), and this algorithm works fine in that case, even with the best performance of the optimized Brute Force Method. i.e.: this degradation of the Brute Force Method has been "hardcoded".